separate helper and test files for better project structure
