+++ tests/unauthenticated/userTest.js
@@ -3,7 +3,7 @@ describe('Unauthenticated: Currently testing user management,', function () {
   var timeRightNow = new Date().getTime()
   var OwnCloud = require('../../src')
   var config = require('../config/config.json')
-  const { getMockServerBaseUrl } = require('../pactHelper.js')
+  const { getMockServerBaseUrl } = require('../helpers/pactHelper.js')
   const mockServerBaseUrl = getMockServerBaseUrl()
 
   // LIBRARY INSTANCE
