+++ tests/unauthorized/groupTest.js
@@ -14,7 +14,7 @@ describe('Unauthorized: Currently testing group management,', function () {
     getCapabilitiesWithInvalidAuthInteraction,
     invalidAuthHeader,
     xmlResponseHeaders
-  } = require('../pactHelper.js')
+  } = require('../helpers/pactHelper.js')
 
   const unauthorizedResponseObject = {
     status: 401,
